Fix crash when removing a single connected signal.
Issue caused by nw (in SetValues()->jlAxis.c:734) having its axis.unit_string set to a non-NULL
but invalid memory location, in the case when SetValues() is called but there is no current signal
being graphed (XtVaSetValues()->StripGraph.c:660 calls SetValues()->jlAxis.c:734).
In this case, we pass the non-NULL check in SetValues() but segfault when strcmp
tries to actually access the string's memory.

Passing "NULL" instead of "0" as the XjNunitString gives us a proper NULL string
inside of SetValues(), which will be correctly handled.